Move ioctl to Windos utils, where it belongs
