+++ qiling/os/utils.py
@@ -9,17 +9,9 @@ This module is intended for general purpose functions that are only used in qili
 
 from typing import Any, MutableMapping, Mapping, Union, Sequence, MutableSequence, Tuple
 from uuid import UUID
-import ctypes
-
-from unicorn import UcError
 
 from qiling import Qiling
 from qiling.const import QL_VERBOSE
-from qiling.os.const import POINTER
-from qiling.os.windows.fncc import STDCALL
-from qiling.os.windows.wdk_const import *
-from qiling.os.windows.structs import *
-from qiling.utils import verify_ret
 
 class QlOsUtils:
 
@@ -213,170 +205,3 @@ class QlOsUtils:
         # we want to rewrite the return address to the function
         self.ql.stack_write(0, start)
 
-    # Emulate DeviceIoControl() of Windows
-    # BOOL DeviceIoControl(
-    #      HANDLE       hDevice,
-    #      DWORD        dwIoControlCode,
-    #      LPVOID       lpInBuffer,
-    #      DWORD        nInBufferSize,
-    #      LPVOID       lpOutBuffer,
-    #      DWORD        nOutBufferSize,
-    #      LPDWORD      lpBytesReturned,
-    #      LPOVERLAPPED lpOverlapped);
-    def ioctl(self, params):
-        heap = self.ql.os.heap
-
-        def ioctl_code(DeviceType, Function, Method, Access):
-            return (DeviceType << 16) | (Access << 14) | (Function << 2) | Method
-
-        alloc_addr = []
-        def build_mdl(buffer_size, data=None):
-            if self.ql.archtype == QL_ARCH.X8664:
-                mdl = MDL64()
-            else:
-                mdl = MDL32()
-
-            mapped_address = heap.alloc(buffer_size)
-            alloc_addr.append(mapped_address)
-            mdl.MappedSystemVa.value = mapped_address
-            mdl.StartVa.value = mapped_address
-            mdl.ByteOffset = 0
-            mdl.ByteCount = buffer_size
-            if data:
-                written = data if len(data) <= buffer_size else data[:buffer_size]
-                self.ql.mem.write(mapped_address, written)
-
-            return mdl
-
-        # quick simple way to manage all alloc memory
-        if self.ql.ostype == QL_OS.WINDOWS:
-            # print("DeviceControl callback is at 0x%x" %self.loader.driver_object.MajorFunction[IRP_MJ_DEVICE_CONTROL])
-            if self.ql.loader.driver_object.MajorFunction[IRP_MJ_DEVICE_CONTROL] == 0:
-                # raise error?
-                return (None, None, None)
-
-            # create new memory region to store input data
-            _ioctl_code, output_buffer_size, in_buffer = params
-            # extract data transfer method
-            devicetype, function, ctl_method, access = _ioctl_code
-
-            input_buffer_size = len(in_buffer)
-            input_buffer_addr = heap.alloc(input_buffer_size)
-            alloc_addr.append(input_buffer_addr)
-            self.ql.mem.write(input_buffer_addr, bytes(in_buffer))
-
-            # create new memory region to store out data
-            output_buffer_addr = heap.alloc(output_buffer_size)
-            alloc_addr.append(output_buffer_addr)
-
-            # allocate memory regions for IRP and IO_STACK_LOCATION
-            if self.ql.archtype == QL_ARCH.X8664:
-                irp_addr = heap.alloc(ctypes.sizeof(IRP64))
-                alloc_addr.append(irp_addr)
-                irpstack_addr = heap.alloc(ctypes.sizeof(IO_STACK_LOCATION64))
-                alloc_addr.append(irpstack_addr)
-                # setup irp stack parameters
-                irpstack = IO_STACK_LOCATION64()
-                # setup IRP structure
-                irp = IRP64()
-                irp.irpstack = ctypes.cast(irpstack_addr, ctypes.POINTER(IO_STACK_LOCATION64))
-            else:
-                irp_addr = heap.alloc(ctypes.sizeof(IRP32))
-                alloc_addr.append(irp_addr)
-                irpstack_addr = heap.alloc(ctypes.sizeof(IO_STACK_LOCATION32))
-                alloc_addr.append(irpstack_addr)
-                # setup irp stack parameters
-                irpstack = IO_STACK_LOCATION32()
-                # setup IRP structure
-                irp = IRP32()
-                irp.irpstack = ctypes.cast(irpstack_addr, ctypes.POINTER(IO_STACK_LOCATION32))
-
-                #print("32 stack location size = 0x%x" %ctypes.sizeof(IO_STACK_LOCATION32))
-                #print("32 status block size = 0x%x" %ctypes.sizeof(IO_STATUS_BLOCK32))
-                #print("32 irp size = 0x%x" %ctypes.sizeof(IRP32))
-                #print("32 IoStatus offset = 0x%x" %IRP32.IoStatus.offset)
-                #print("32 UserIosb offset = 0x%x" %IRP32.UserIosb.offset)
-                #print("32 UserEvent offset = 0x%x" %IRP32.UserEvent.offset)
-                #print("32 UserBuffer offset = 0x%x" %IRP32.UserBuffer.offset)
-                #print("32 irpstack offset = 0x%x" %IRP32.irpstack.offset)
-                #print("irp at %x, irpstack at %x" %(irp_addr, irpstack_addr))
-
-            self.ql.log.info("IRP is at 0x%x, IO_STACK_LOCATION is at 0x%x" %(irp_addr, irpstack_addr))
-
-            irpstack.Parameters.DeviceIoControl.IoControlCode = ioctl_code(devicetype, function, ctl_method, access)
-            irpstack.Parameters.DeviceIoControl.OutputBufferLength = output_buffer_size
-            irpstack.Parameters.DeviceIoControl.InputBufferLength = input_buffer_size
-            irpstack.Parameters.DeviceIoControl.Type3InputBuffer.value = input_buffer_addr # used by IOCTL_METHOD_NEITHER
-            self.ql.mem.write(irpstack_addr, bytes(irpstack))
-
-            if ctl_method == METHOD_NEITHER:
-                irp.UserBuffer.value = output_buffer_addr  # used by IOCTL_METHOD_NEITHER
-
-            # allocate memory for AssociatedIrp.SystemBuffer
-            # used by IOCTL_METHOD_IN_DIRECT, IOCTL_METHOD_OUT_DIRECT and IOCTL_METHOD_BUFFERED
-            system_buffer_size = max(input_buffer_size, output_buffer_size)
-            system_buffer_addr = heap.alloc(system_buffer_size)
-            alloc_addr.append(system_buffer_addr)
-
-            # init data from input buffer
-            self.ql.mem.write(system_buffer_addr, bytes(in_buffer))
-            irp.AssociatedIrp.SystemBuffer.value = system_buffer_addr
-
-            if ctl_method in (METHOD_IN_DIRECT, METHOD_OUT_DIRECT):
-                # Create MDL structure for output data
-                # used by both IOCTL_METHOD_IN_DIRECT and IOCTL_METHOD_OUT_DIRECT
-                mdl = build_mdl(output_buffer_size)
-                if self.ql.archtype == QL_ARCH.X8664:
-                    mdl_addr = heap.alloc(ctypes.sizeof(MDL64))
-                else:
-                    mdl_addr = heap.alloc(ctypes.sizeof(MDL32))
-
-                alloc_addr.append(mdl_addr)
-
-                self.ql.mem.write(mdl_addr, bytes(mdl))
-                irp.MdlAddress.value = mdl_addr
-
-            # everything is done! Write IRP to memory
-            self.ql.mem.write(irp_addr, bytes(irp))
-
-            # set function args
-            self.ql.log.info("Executing IOCTL with DeviceObject = 0x%x, IRP = 0x%x" %(self.ql.loader.driver_object.DeviceObject, irp_addr))
-            # TODO: make sure this is indeed STDCALL
-            self.ql.os.fcall = self.ql.os.fcall_select(STDCALL)
-            self.ql.os.fcall.writeParams(((POINTER, self.ql.loader.driver_object.DeviceObject), (POINTER, irp_addr)))
-
-            try:
-                # now emulate IOCTL's DeviceControl
-                self.ql.run(self.ql.loader.driver_object.MajorFunction[IRP_MJ_DEVICE_CONTROL])
-            except UcError as err:
-                verify_ret(self.ql, err)
-
-            # read current IRP state
-            if self.ql.archtype == QL_ARCH.X8664:
-                irp_buffer = self.ql.mem.read(irp_addr, ctypes.sizeof(IRP64))
-                irp = IRP64.from_buffer(irp_buffer)
-            else:
-                irp_buffer = self.ql.mem.read(irp_addr, ctypes.sizeof(IRP32))
-                irp = IRP32.from_buffer(irp_buffer)
-
-            io_status = irp.IoStatus
-
-            # read output data
-            output_data = b''
-            if io_status.Status.Status >= 0:
-                if ctl_method == METHOD_BUFFERED:
-                    output_data = self.ql.mem.read(system_buffer_addr, io_status.Information.value)
-                if ctl_method in (METHOD_IN_DIRECT, METHOD_OUT_DIRECT):
-                    output_data = self.ql.mem.read(mdl.MappedSystemVa.value, io_status.Information.value)
-                if ctl_method == METHOD_NEITHER:
-                    output_data = self.ql.mem.read(output_buffer_addr, io_status.Information.value)
-
-            # now free all alloc memory
-            for addr in alloc_addr:
-                # print("freeing heap memory at 0x%x" %addr) # FIXME: the output is not deterministic??
-                heap.free(addr)
-            #print("\n")
-
-            return io_status.Status.Status, io_status.Information.value, output_data
-        else: # TODO: IOCTL for non-Windows.
-            pass
