+++ tests/test_pe_sys.py
@@ -259,7 +259,7 @@ class PETest(unittest.TestCase):
         IOCTL_SIOCTL_METHOD_OUT_DIRECT = (40000, 0x901, METHOD_OUT_DIRECT, FILE_ANY_ACCESS)
         output_buffer_size = 0x1000
         in_buffer = b'Test input\0'
-        Status, Information_value, output_data = ql.os.utils.ioctl((IOCTL_SIOCTL_METHOD_OUT_DIRECT, output_buffer_size, in_buffer))
+        Status, Information_value, output_data = utils.ioctl(ql, (IOCTL_SIOCTL_METHOD_OUT_DIRECT, output_buffer_size, in_buffer))
 
         expected_result = b'This String is from Device Driver !!!\x00'
         self.assertEqual(Status, 0)
